Testing code for issue #36
For some reason, the periodicEvents ArrayList changes when referenced
through getPeriodicEvents() outside of the class

In this commit are two print statements that are called when trying to
schedule a periodic event: one inside the Scheduler, and one in
MainServer after the event has been scheduled. They should both print
the same thing, however non-daily periodic events only print the latest
day they are scheduled for in the week, starting from Sunday.